app-update - remaining 5.1 default configurations
Most of the rest of the differences in the default
app:update rake task are handled in this commit,
including changes in:

* application.rb
* environment.rb
* environments files

This also handles a bug which is evidently fixed somewhere
in the Rails source, apparently not for the particular
revision here.

The `listen` gem was installed as well.