update to ruby 2.4.2
This seems like a lame commit, and in some ways it is,
but updating a ruby version still takes time some time
and attention to rebuild the gems, run cucumber and rspec
and in general make sure nothing is screwed up. So in
that sense, a very small commit is a very large success.

Updating where a lot of code needs to be changed is
accordingly not so much a success. Why should an
application break between two minor revisions in a
programming language? Maybe a smell there.