controller-testing - gem removed, all specs pass
Regardless of title of this commit, this is just the
first pass at cleaning up controller specs. The idea
is to take the existing controller specs and split
them out into controller, request and feature specs.

Splitting everything will result in a significantly
overtested code, but it's useful to me for comparing
various techniques.

Also, the current request specs are in need of some
considerable refactoring, but that will probably wait
until after the split.

If this were production code, I'd probably do just some
request spec refactoring and submit PR as-is. It's too
easy to make a career out of testing.