devise-token-auth - working request spec
As it turns out it's easy to write request specs
by specifying the header values in the request
itself rather then merging it into the request.headers
attribute. Apparently I just needed a week to let it
simmer in my head and connect the dots.

Will likely add more to this branch before bringing
it back into master, profiles need to be added.