ruby 2.5.0 - update with some gems
There was some mankiness related to devise
updating, everything seemed to go fine.

Initial commit message follows:

This commit is a placeholder for when devise_token_auth
is updated to work with devise 4.4.0. The problem
is a ruby syntax error in devise < 4.4.0.

Can't go any further until devise_token_auth is
updated.