TASKLETS-10 count descendants with CTE SQL query
Rails doesn't support CTE queries, so it's necessary
to construct CTEs using SQL then executing the query
via a database connection.

The query returns a tuple of hashes, which isn't terribly
useful if one is interested in either the actual models,
or representing the nested structure.

It's trivial to count the results which is what this
change demonstrates.