TASKLETS-13 count children correctly.
Finding a tree by its parent id is a problem when there
is more than one tree, and the tree is rooted by having
a parent id equal to null. It makes it difficult to
query just a single tree.

In this change, every node is labeled, including the root
node. For now, labels are not unique, but that will need to
be dealt with at some point.