TASKLETS-11 create tree from cte results
The CTE query is currently returning a tuple of the selected
attribute, which is to say, an array of hashes. This is faster
than making multiple database calls.

Once we have the results of the CTE query, those results need
to be processed into a tree form, which this commit is doing
in software.

The hardest part of this work is figuring out how to test the
results, which are nested hashes. Order matters. There is
opportunity for future improvement here to be sure.