TASKLETS-34 implement Task destroy
Destroy can be implemented in several ways. The easiest
way is to denote the association on tasks with `dependentent:destroy`.
However, as shown by a log fragment written in a comment in the
task_spec, Rails makes a database call for each record being
destroyed. That's expensive.

Foreign key alternative:

CTE alternative: