TASKLETS-75 move CTE query builder to service class
Common Table Expressions are not yet handled natively in Rails,
so here is a service class for building such queries. The class
does not do any database access at, it simply builds a SQL statement
which is handed off to the caller.

The methods in the class were extracted with the class from the Task
model.

The purpose of testing the CteQueryBuilder is to make it more difficult
to introduce regressions into the code. The SQL can be very sensitive
and if it fails, difficult to debug. The specs here enforce consistency
between the class and the specs, and have nothing to do with the validity
of the SQL, which is tested elsewhere.