TASKLETS-57 set up sentry
This took me a lot longer than it should have. I
got sidetracked by an improperly formatted redis
uri: it needs to start with 'redis://' instead of
'http://'.

Then was sidetracked into configuring processor
for custom sanitizer.

In both cases had to go read the source code, which
is actually a win.