TASKLETS-133 add webpacker capability
https://doolin.atlassian.net/browse/TASKLETS-133

This concludes my Rails 6 upgrade. I'm using the split method
where sprockets handles everything but Javascript, and webpacker
handles all of the Javascript. It might be worth removing
sprockets completely at some point in the future, but this is
fine for now. The node packages will need to be kept up to date
as well as the ruby gems.