TASKLETS-159 swap children between two tasks
Implementing `update` depends on what, exactly is
being updated. Updating the structure of the tree
is distinctly different than updating a record in
the database.

`swap_children` does what it says: takes two nodes
and swaps the children from one to another.

The Rails implementation is interesting as the
local state in the application serves as the `temp`
variable in a swapping routine. Updating a Rails
collection proxy won't work, one task ends up with
all the children.